Let tools.deps build the classpath
This still uses :paths to set the :resource-paths as in earlier releases, but now it uses :source-paths for not only :extra-paths but also for any directories added to classpath by :git or :local deps.

In addition, the dependencies are no longer handed to Boot to (re)load. Instead they are added directly to Boot's classpath (since tools.deps has done all the work).

TODO: update the README!